Move Home components to Home dir
